Fix validation & remove unused codes
